Refactored to match changes in core package.
